+++ bugsnag-android-core/src/main/java/com/bugsnag/android/Configuration.java
@@ -4,6 +4,7 @@ import android.content.Context;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.annotation.VisibleForTesting;
 
 import java.io.File;
 import java.util.Locale;
@@ -19,7 +20,7 @@ public class Configuration implements CallbackAware, MetadataAware, UserAware {
 
     private static final int MIN_BREADCRUMBS = 0;
     private static final int MAX_BREADCRUMBS = 100;
-    private static final String API_KEY_REGEX = "[A-Fa-f0-9]{32}";
+    private static final int VALID_API_KEY_LEN = 32;
     private static final long MIN_LAUNCH_CRASH_THRESHOLD_MS = 0;
 
     final ConfigInternal impl;
@@ -47,14 +48,29 @@ public class Configuration implements CallbackAware, MetadataAware, UserAware {
     }
 
     private void validateApiKey(String value) {
-        if (Intrinsics.isEmpty(value)) {
+        if (isInvalidApiKey(value)) {
+            DebugLogger.INSTANCE.w("Invalid configuration. "
+                    + "apiKey should be a 32-character hexademical string, got " + value);
+        }
+    }
+
+    @VisibleForTesting
+    static boolean isInvalidApiKey(String apiKey) {
+        if (Intrinsics.isEmpty(apiKey)) {
             throw new IllegalArgumentException("No Bugsnag API Key set");
         }
-
-        if (!value.matches(API_KEY_REGEX)) {
-            DebugLogger.INSTANCE.w(String.format("Invalid configuration. apiKey should be a "
-                    + "32-character hexademical string, got \"%s\"", value));
+        if (apiKey.length() != VALID_API_KEY_LEN) {
+            return true;
         }
+        // check whether each character is hexadecimal (either a digit or a-f).
+        // this avoids using a regex to improve startup performance.
+        for (int k = 0; k < VALID_API_KEY_LEN; k++) {
+            char chr = apiKey.charAt(k);
+            if (!Character.isDigit(chr) && (chr < 'a' || chr > 'f')) {
+                return true;
+            }
+        }
+        return false;
     }
 
     private void logNull(String property) {